ai/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-medical-ai"
version = "0.1.0"
description = "AI module for medication adherence prediction and analytics."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116,<1.0",
  "uvicorn>=0.35,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "pandas>=2.0,<3.0",
  "numpy>=1.24,<3.0",
  "scikit-learn>=1.6,<2.0",
  "joblib>=1.4,<2.0",
  "langchain-core>=0.3,<1.0",
  "langchain-google-genai>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.28,<1.0",
]
xgboost = [
  "xgboost>=2.1,<3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
